#RegionNotifier: split region job into RegionFilter
