+++ src/main/java/org/altbeacon/beacon/notifier/RegionRangeNotifier.java
@@ -13,45 +13,25 @@ import java.util.List;
  */
 public abstract class RegionRangeNotifier implements RangeNotifier{
 
-    private final List<Region> regions = new ArrayList<>();
+    private final RegionFilter regionFilter = new RegionFilter();
 
-    /**
-     * @param region
-     */
-    public void addRegion(Region region){
-        synchronized (regions){
-            regions.add(region);
-        }
+
+    public void addRegion(Region region) {
+        regionFilter.addRegion(region);
     }
 
-    /**
-     * @param region
-     * @return
-     */
-    public boolean removeRegion(Region region){
-        synchronized (regions) {
-            return regions.remove(region);
-        }
+
+    public boolean removeRegion(Region region) {
+        return regionFilter.removeRegion(region);
     }
 
-    /**
-     * Remove all regions
-     */
-    public void removeAllRegions(){
-        synchronized (regions){
-            regions.clear();
-        }
-    }
-
-    private boolean containRegion(Region region){
-        synchronized (regions){
-            return regions.contains(region);
-        }
+    public void removeAllRegions() {
+        regionFilter.removeAllRegions();
     }
 
     @Override
     public void didRangeBeaconsInRegion(Collection<Beacon> beacons, Region region) {
-        if(containRegion(region)){
+        if(regionFilter.containRegion(region)){
             didRangeBeaconsInReferencedRegion(beacons, region);
         }
     }
